ub boards check prototype
UB boards checking first realization. Try to model and check how UEKV will proccess the UB data. Set the voltage levels on UB connector by wires. With manual change of vol. levels the indications works correctly. But the rgb leds order should to be modified as its is on the board.